% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphpdd.R
\name{graphpdd}
\alias{graphpdd}
\title{Une fonction permettant de produire des graphiques pour le menu PDD.}
\usage{
graphpdd(data, type_general, lib_var, weight = NULL, title = "")
}
\arguments{
\item{data}{une data frame.

- S'il s'agit d'une \bold{analyse monovariée} et que le type général est \bold{Qualitatif}, \bold{Echelle sémantique} ou \bold{Echelle sémantique inversée} : le nombre de lignes de \code{data} est égal au nombre de modalités de la variable à analyser et le nombre de colonne de \code{data} vaut 1. (NB : dans ce cas \code{rownames(data)} doit contenir les noms de modalités).

- S'il s'agit d'une \bold{analyse monovariée} et que le type général est \bold{Quantitatif}.  : le nombre de lignes de \code{data} est égal au nombre d'observations de la population à analyser et le nombre de colonne \code{data} vaut 1.

- S'il s'agit d'une \bold{analyse bivariée} et que les deux types généraux sont \bold{Qualitatif}, \bold{Echelle sémantique} ou \bold{Echelle sémantique inversée} : le nombre de lignes de \code{data} est égal au nombre de modalités de la première variable à analyser et le nombre de colonne de \code{data} au nombre de modalités de la seconde variable à analyser. (NB : dans ce cas \code{rownames(data)} doit contenir les noms des modalités de la première variable et \code{colnames(data)} ceux de la seconde variable).

- S'il s'agit d'une \bold{analyse bivariée} et que les deux types généraux sont \bold{Quantitatif} : le nombre de lignes de \code{data} est égal au nombre d'observation de la population à analyser et le nombre de colonnes \code{data} vaut 2.

- S'il s'agit d'une \bold{analyse bivariée} et le premier type général est \bold{Quantitatif} et le second \bold{Qualitatif}, \bold{Echelle sémantique} ou \bold{Echelle sémantique inversée} : le nombre de lignes de \code{data} est égal au nombre d'observations de la population à analyser et le nombre de colonne \code{data} vaut 2 ; la première colonne correspondant aux données quantitatives, et la seconde aux données qualitatives (ou sémantiques).

- S'il s'agit d'une \bold{analyse bivariée} et le premier type général est \bold{Qualitatif}, \bold{Echelle sémantique} ou \bold{Echelle sémantique inversée} et le second \bold{Quantitatif} : le nombre de lignes de \code{data} est égal au nombre d'observations de la population à analyser et le nombre de colonne \code{data} vaut 2 ; la première colonne correspondant aux données qualitatives (ou sémantiques) et la seconde aux données quantitatives.}

\item{type_general}{un vecteur de une ou deux chaines de caractères parmis \code{c("Quantitatif","Qualitatif","Echelle sémantique","Echelle sémantique inversée")}.}

\item{lib_var}{un vecteur de une ou deux chaines de caractères indiquant le ou les libellés de la (des) variable(s) à représenter.}

\item{weight}{\code{NULL} ou un vecteur de poids de longueur égale au nombre d'observation totale (utile que si au moins une des variables à analyser est \bold{Quantitatif} et que l'analyse est pondérée).}

\item{title}{une chaine de caractère représentant le titre de l'analyse.}
}
\description{
Les analyses dépendent du fait qu'il y ait une ou deux variables à analyser ainsi que du type (général) de chacune des variables à analyser.
}
\examples{
######################################
# Exemple mono qualitatif
######################################
data<-data.frame(c(10000,2000,888))
rownames(data)<-c("Un premier nom de modalité","Un deuxième nom de modalité","Un troisième nom de modalité")
p <- graphpdd(data = data
              ,lib_var = "Un libellé de variable"
              ,type_general = "Qualitatif"
              ,title = "Titre du graphique")
print(p)
######################################
# Exemple mono quantitatif
######################################
set.seed(123)
x<-runif(1000)
data<-data.frame(x=rnorm(n=1000,mean = 0,sd=1)*(x>0.5)+(x<=0.5)*rnorm(n=1000,mean=5,sd=2))
p <- graphpdd(data = data
              ,lib_var = "Un libellé de variable"
              ,type_general = "Quantitatif"
              ,title = "Titre du graphique")
print(p)

######################################
# Exemple bivarié quali x quali
######################################
set.seed(123)
a<-paste0("Le libellé de la modalité (var1) numéro ",seq(7))
b<-paste0("Le libellé de la modalité (var2) numéro ",seq(9))
data<-reshape2::dcast(data.frame(x=sample(a,1000,replace=TRUE),y=sample(b,1000,replace=TRUE)),x~y)
rownames(data)<-data[,1]
data<-data[,-1]
lib_var<-
  type_general<-c("Qualitatif","Qualitatif")
title="Un titre"

p <- graphpdd(data = data
              ,lib_var = c("Une première variable avec un libellé super long","Une seconde variable")
              ,type_general = c("Qualitatif","Qualitatif")
              ,title = "Titre du graphique")
print(p)

p <- graphpdd(data = data
              ,lib_var = c("Une première variable avec un libellé super long","Une seconde variable")
              ,type_general = c("Qualitatif","Echelle sémantique")
              ,title = "Titre du graphique")
print(p)


######################################
# Exemple  bivarié quanti x quanti
######################################
set.seed(123)
x<-runif(10000)
x<-rnorm(n=10000,mean = 0,sd=1)*(x>0.5)+(x<=0.5)*rnorm(n=10000,mean=5,sd=2)
y<-x+0.25*x^2+2+rnorm(n=10000,mean=0,sd=3)
data<-data.frame(x=x,y=y)
p <- graphpdd(data = data
              ,lib_var = c("Une première variable avec un libellé super long","Une seconde variable")
              ,type_general = c("Quantitatif","Quantitatif")
              ,title = "Titre du graphique")
print(p)
# avec un poids
w<-log(1+abs(data[,1]))*(1+runif(10000)/5)
p <- graphpdd(data = data
              ,lib_var = c("Une première variable avec un libellé super long","Une seconde variable")
              ,type_general = c("Quantitatif","Quantitatif")
              , weight = w
              ,title = "Titre du graphique (pondéré)")
print(p)

######################################
# Exemple  bivarié quanti x quali
######################################
set.seed(123)
x<-runif(10000)
x<-rnorm(n=10000,mean = 0,sd=1)*(x>0.5)+(x<=0.5)*rnorm(n=10000,mean=5,sd=2)
b<-paste0("Le libellé de la modalité (var2) numéro ",seq(6))
y<-sample(b,10000,replace=TRUE)
data<-data.frame(x=x,y=y)

p <- graphpdd(data = data
              ,lib_var = c("Une première variable avec un libellé super long","Une seconde variable")
              ,type_general = c("Quantitatif","Qualitatif")
              ,title = "Titre du graphique")
print(p)
# On peut inverser les données (quanti <-> quali), cela ne change rien
p <- graphpdd(data = data[,2:1]
              ,lib_var = c("Une première variable avec un libellé super long","Une seconde variable")
              ,type_general = c("Qualitatif","Quantitatif")
              ,title = "Titre du graphique (2)")
print(p)


}

